* add-support-for-small-armature:
  Fix various compilation warnings
  Restore dq interpolation for armature with only 2 bones

+++ Modules/CLI/EvalSurfaceWeight/EvalSurfaceWeight.cxx
@@ -273,7 +273,7 @@ int main( int argc, char * argv[] )
     }
 
   if (sampleVertices.size() <= 0
-    || sampleVertices.size() != points->GetNumberOfPoints())
+      || sampleVertices.size() != static_cast<size_t>(points->GetNumberOfPoints()))
     {
     std::cout<<"FAILURE: "<<sampleVertices.size()<<" out of "
       <<points->GetNumberOfPoints()
